+++ app/models/cms/user_group_membership.rb
@@ -4,6 +4,6 @@ module Cms
     extend Cms::DefaultAccessible
 
     belongs_to :group, :class_name => 'Cms::Group'
-    belongs_to :user, :class_name => 'Cms::User'
+    belongs_to :user, :class_name => 'Cms::PersistentUser'
   end
 end