+++ db/migrate/20131218222005_create_cms_external_users.rb
@@ -1,0 +1,10 @@
+class CreateCmsExternalUsers < ActiveRecord::Migration
+  def change
+    change_table :cms_users do |t|
+      t.column :type, :string, default: 'Cms::User'
+      t.column :source, :string
+      t.text :external_data
+    end
+
+  end
+end