+++ spec/cms/authentication/test_password_strategy.rb
@@ -1,0 +1,9 @@
+require "minitest_helper"
+
+describe Cms::Authencation::TestPasswordStrategy do
+  describe 'method-name' do
+    it 'should ' do
+
+    end
+  end
+end